+++ KeePassium AutoFill/MainCoordinator.swift
@@ -590,6 +590,7 @@ extension MainCoordinator: WatchdogDelegate {
         let context = LAContext()
         let policy = LAPolicy.deviceOwnerAuthenticationWithBiometrics
         context.localizedFallbackTitle = "" // hide "Enter Password" fallback; nil won't work
+        context.localizedCancelTitle = LString.actionUsePasscode
         
         Diag.debug("Biometric auth: showing request")
         context.evaluatePolicy(policy, localizedReason: LString.titleTouchID) {