+++ KeePassium/util/LString.swift
@@ -31,7 +31,7 @@ public enum LString {
     public static let actionRemoveFile = NSLocalizedString("Remove", comment: "Action/button to remove file from the app (the file remains, but the app forgets about it)")
     public static let actionUnlock = NSLocalizedString("Unlock", comment: "Action/button to unlock the App Lock with passcode")
     public static let actionContactUs = NSLocalizedString("Contact Us", comment: "Action/button to write an email to support")
-    
+    public static let actionUsePasscode = "Use Passcode".localized(comment: "Action/button to switch from TouchID/FaceID prompt to manual input of the AppLock passcode.")
 
     public static let titleError = NSLocalizedString("Error", comment: "Title of an error message notification")
     public static let titleWarning = NSLocalizedString("Warning", comment: "Title of an warning message")