+++ KeePassiumLib/KeePassiumLib/db/totp/TOTPGenerator.swift
@@ -8,140 +8,6 @@
 
 import Foundation
 
-open class TOTPGeneratorFactory {
-    enum SplitFormat {
-        public static let seedFieldName = "TOTP Seed"
-        public static let settingsFieldName = "TOTP Settings"
-    }
-    
-    enum URIFormat {
-        public static let fieldName = "otp"
-        
-        public static let scheme = "otpauth"
-        public static let host = "totp"
-        
-        public static let seedParam = "secret"
-        public static let timeStepParam = "period"
-        public static let lengthParam = "digits"
-        public static let algorithmParam = "algorithm"
-        
-        public static let defaultTimeStep = 30
-        public static let defaultLength = 6
-        public static let defaultAlgorithm = "SHA1"
-    }
-    
-    public static func makeGenerator(for entry: Entry) -> TOTPGenerator? {
-        return makeGenerator(from: entry.fields)
-    }
-    
-    public static func makeGenerator(from fields: [EntryField]) -> TOTPGenerator? {
-        if let totpURIField = fields.first(where: { $0.name == URIFormat.fieldName }) {
-            return TOTPGeneratorFactory.makeGenerator(uri: totpURIField.value)
-        } else {
-            guard let seedField =
-                fields.first(where: { $0.name == SplitFormat.seedFieldName })
-                else { return nil }
-            guard let settingsField =
-                fields.first(where: { $0.name == SplitFormat.settingsFieldName })
-                else { return nil }
-            return TOTPGeneratorFactory.makeGenerator(
-                seed: seedField.value,
-                settings: settingsField.value)
-        }
-    }
-    
-    public static func makeGenerator(uri uriString: String) -> TOTPGenerator? {
-         guard let components = URLComponents(string: uriString),
-            components.scheme == URIFormat.scheme,
-            components.host == URIFormat.host,
-            let queryItems = components.queryItems else
-        {
-            Diag.warning("OTP URI has unexpected format")
-            return nil
-        }
-        
-        let params = queryItems.reduce(into: [String: String]()) { (result, item) in
-            result[item.name] = item.value
-        }
-        
-        guard let seedString = params[URIFormat.seedParam],
-            let seedData = base32DecodeToData(seedString),
-            !seedData.isEmpty else
-        {
-            Diag.warning("OTP parameter cannot be parsed [parameter: \(URIFormat.seedParam)]")
-            return nil
-        }
-        
-        if let algorithm = params[URIFormat.algorithmParam],
-            algorithm != URIFormat.defaultAlgorithm
-        {
-            Diag.warning("OTP algorithm is not supported [algorithm: \(algorithm)]")
-            return nil
-        }
-        guard let timeStep = Int(params[URIFormat.timeStepParam] ?? "\(URIFormat.defaultTimeStep)") else {
-            Diag.warning("OTP parameter cannot be parsed [parameter: \(URIFormat.timeStepParam)]")
-            return nil
-        }
-        guard let length = Int(params[URIFormat.lengthParam] ?? "\(URIFormat.defaultLength)") else {
-            Diag.warning("OTP parameter cannot be parsed [parameter: \(URIFormat.lengthParam)]")
-            return nil
-        }
-        
-        return TOTPGeneratorRFC6238(
-            seed: ByteArray(data: seedData),
-            timeStep: timeStep,
-            length: length)
-    }
-    
-    public static func makeGenerator(
-        seed seedString: String,
-        settings settingsString: String
-    ) -> TOTPGenerator? {
-        guard let seed = parseSeedString(seedString) else {
-            Diag.warning("Unrecognized TOTP seed format")
-            return nil
-        }
-        
-        let settings = settingsString.split(separator: ";")
-        guard settings.count == 2 else {
-            Diag.warning("Unexpected TOTP settings number [expected: 2, got: \(settings.count)]")
-            return nil
-        }
-        guard let timeStep = Int(settings[0]) else {
-            Diag.warning("Failed to parse TOTP time step as Int")
-            return nil
-        }
-        guard timeStep > 0 else {
-            Diag.warning("Invalid TOTP time step value: \(timeStep)")
-            return nil
-        }
-        
-        if let length = Int(settings[1]) {
-            return TOTPGeneratorRFC6238(seed: seed, timeStep: timeStep, length: length)
-        } else if settings[1] == TOTPGeneratorSteam.typeSymbol {
-            return TOTPGeneratorSteam(seed: seed, timeStep: timeStep)
-        } else {
-            Diag.warning("Unexpected TOTP size or type: '\(settings[1])'")
-            return nil
-        }
-    }
-    
-    static func parseSeedString(_ seedString: String) -> ByteArray? {
-        let cleanedSeedString = seedString
-            .replacingOccurrences(of: " ", with: "")
-            .replacingOccurrences(of: "=", with: "")
-        if let seedData = base32DecodeToData(cleanedSeedString) {
-            return ByteArray(data: seedData)
-        }
-        if let seedData = base32HexDecodeToData(cleanedSeedString) {
-            return ByteArray(data: seedData)
-        }
-        if let seedData = Data(base64Encoded: cleanedSeedString) {
-            return ByteArray(data: seedData)
-        }
-        return nil
-    }
-}
 
 public protocol TOTPGenerator: class {
     var elapsedTimeFraction: Float { get }
@@ -177,8 +43,9 @@ public class TOTPGeneratorRFC6238: TOTPGenerator {
     
     public var elapsedTimeFraction: Float { return getElapsedTimeFraction(timeStep: timeStep) }
 
-    fileprivate init?(seed: ByteArray, timeStep: Int, length: Int) {
+    internal init?(seed: ByteArray, timeStep: Int, length: Int) {
         guard length >= 4 && length <= 8 else { return nil }
+        guard timeStep > 0 else { return nil }
         
         self.seed = seed
         self.timeStep = timeStep
@@ -208,7 +75,9 @@ public class TOTPGeneratorSteam: TOTPGenerator {
     private let timeStep: Int
     private let length = 5
 
-    fileprivate init?(seed: ByteArray, timeStep: Int) {
+    internal init?(seed: ByteArray, timeStep: Int) {
+        guard timeStep > 0 else { return nil }
+        
         self.seed = seed
         self.timeStep = timeStep
     }